Adjusted nav files
Organized the Nav files so that they're less clustered
