+++ focusfeed/lib/features/screens/nav/nav_tab.dart
@@ -1,0 +1,1 @@
+enum NavTab { feed, saved, library, settings }
